requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.7
Werkzeug>=3.0.0
